check file doesn't already exist on TM

+++ sandbox/request_processors.py
@@ -1,4 +1,5 @@
 import io
+import os
 
 def receive_data(s):
     # flag denoting when stream of image data begins
@@ -51,8 +52,13 @@ def process_image_data(image_data, filename):
     # TODO: naming convention and directory for images
     # possibly just name after question?
     # could images just be stored on TM?
-    with open(filename, "wb") as f:
-        f.write(image_data)
+
+    # check file hasn't already been requested
+    if not os.path.exists(filename):
+        with open(filename, "wb") as f:
+            f.write(image_data)
+    else:
+        print("file already exists on TM!")
 
 # reads stream of q_ids and stores in list
 def process_generated_questions(qb_response):
